refactor/godClass-javaClassUtil
